[chore] #155 Change package home name to member
