+++ data/src/main/java/com/cmc/curtaincall/data/source/remote/MemberRemoteSource.kt
@@ -4,9 +4,9 @@ import com.cmc.curtaincall.core.network.service.member.MemberService
 import com.cmc.curtaincall.core.network.service.member.request.DeleteMemberRequest
 import com.cmc.curtaincall.core.network.service.member.request.MemberCreateRequest
 import com.cmc.curtaincall.core.network.service.member.request.UpdateMemberRequest
-import com.cmc.curtaincall.domain.model.home.MemberInfoModel
-import com.cmc.curtaincall.domain.model.home.MyParticipationModel
-import com.cmc.curtaincall.domain.model.home.MyRecruitmentModel
+import com.cmc.curtaincall.domain.model.member.MemberInfoModel
+import com.cmc.curtaincall.domain.model.member.MyParticipationModel
+import com.cmc.curtaincall.domain.model.member.MyRecruitmentModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
