+++ domain/src/main/java/com/cmc/curtaincall/domain/model/member/MyParticipationModel.kt
@@ -1,19 +1,19 @@
-package com.cmc.curtaincall.domain.model.home
+package com.cmc.curtaincall.domain.model.member
 
-data class MyRecruitmentModel(
+data class MyParticipationModel(
     val id: Int = Int.MIN_VALUE,
     val title: String = "",
     val curMemberNum: Int = 0,
     val maxMemberNum: Int = 0,
-    val showAt: String? = null,
+    val showAt: String = "",
     val createdAt: String = "",
     val category: String = "",
     val creatorId: Int = 0,
     val creatorNickname: String = "",
     val creatorImageUrl: String? = null,
-    val showId: String? = null,
-    val showName: String? = null,
-    val showPoster: String? = null,
-    val facilityId: String? = null,
-    val facilityName: String? = null
+    val showId: String = "",
+    val showName: String = "",
+    val showPoster: String = "",
+    val facilityId: String = "",
+    val facilityName: String = ""
 )
