+++ domain/src/main/java/com/cmc/curtaincall/domain/model/member/MyRecruitmentModel.kt
@@ -1,19 +1,19 @@
-package com.cmc.curtaincall.domain.model.home
+package com.cmc.curtaincall.domain.model.member
 
-data class MyParticipationModel(
+data class MyRecruitmentModel(
     val id: Int = Int.MIN_VALUE,
     val title: String = "",
     val curMemberNum: Int = 0,
     val maxMemberNum: Int = 0,
-    val showAt: String = "",
+    val showAt: String? = null,
     val createdAt: String = "",
     val category: String = "",
     val creatorId: Int = 0,
     val creatorNickname: String = "",
     val creatorImageUrl: String? = null,
-    val showId: String = "",
-    val showName: String = "",
-    val showPoster: String = "",
-    val facilityId: String = "",
-    val facilityName: String = ""
+    val showId: String? = null,
+    val showName: String? = null,
+    val showPoster: String? = null,
+    val facilityId: String? = null,
+    val facilityName: String? = null
 )
