+++ feature/mypage/src/main/java/com/cmc/curtaincall/feature/mypage/MyPageEvent.kt
@@ -1,7 +1,7 @@
 package com.cmc.curtaincall.feature.mypage
 
 import com.cmc.curtaincall.core.base.BaseEvent
-import com.cmc.curtaincall.domain.model.home.MemberInfoModel
+import com.cmc.curtaincall.domain.model.member.MemberInfoModel
 
 sealed class MyPageEvent : BaseEvent {
     data class LoadMemberInfo(
