+++ feature/mypage/src/main/java/com/cmc/curtaincall/feature/mypage/MyPageUiState.kt
@@ -1,7 +1,7 @@
 package com.cmc.curtaincall.feature.mypage
 
 import com.cmc.curtaincall.core.base.BaseState
-import com.cmc.curtaincall.domain.model.home.MemberInfoModel
+import com.cmc.curtaincall.domain.model.member.MemberInfoModel
 
 data class MyPageUiState(
     val memberInfoModel: MemberInfoModel = MemberInfoModel()
