+++ feature/mypage/src/main/java/com/cmc/curtaincall/feature/mypage/write/MemberViewModel.kt
@@ -1,0 +1,7 @@
+package com.cmc.curtaincall.feature.mypage.write
+
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class MemberViewModel @Inject constructor()
